Switch the DB migration method form instructions to database-migrate.txt, to numbered update scripts. To install updates:
	boss> cd /your/srcdir/sql/updates
	boss> perl /your/objdir/db/dbupdate tbdb

For developers, please look in the sql/updates/4/1 file for an example
of how to do updates. For each revision in database-create.sql, create
a corresponding script and add it to the repository. The scripts will
be applied in numeric order (see db/dbupdate.in), starting with the
version recorded in the version_info table. This table is updated by
dbupdate as it runs each script.