Use this script to delete slices from any of the three DBs:
	Usage: cleanupslice [-f] -c | -m | -s <uuid>

where -c is clearinghouse, -m is the component manager, and -s is the
slice authority. Note that -c only makes sense on Utah Emulab since
that is where the clearinghouse is. When using the -s option, the
slice is cleared from the local SA and the clearinghouse by issuing an
RPC call to remove it. For -m, the slice and its slivers are
deallocated.

This script will eventuallt get called from something to handle slice
expiration on the clearinghouse and the slice authority (the current
expire daemon only deals the CM since that is resources get tied up,
so its most important).

We have yet to figure out an approach for expiration at the SA and the
clearinghouse. Never did like that slice renew thing.