Add two calls to CM interface; SliverStatus() and SliceStatus() which allow you to find out the status of your sliver. As you can guess, one takes a sliver credential and the other takes the slice credential. See the test scripts test/slicestatus.py and test/slicestatus.py. Both calls return the same thing. In pythonese:
{'status': 'notready',
 'details': {'de9803c2-773e-102b-8eb4-001143e453fe': 'notready'}
}

The 'status' is a summary of the entire aggregate; if all nodes are
ISUP then the status is 'ready'

The details is an array of the status of each sliver, indexed by the
sliver uuid.

Probably need to define a better set of status indicators, but this is
the one that was bothering me the most.