Add code to os_setup to log information about Image usage. Enough information is logged so that, at any point in time, it is possible to tell what images are being used.  After collecting some stats for a while I hope to use this data to evaluate various strategies for preloading disks with images other than the default.
Although not its primary purpose, enough information is
collection to be able to get a snapshot of node usage at any
point in time.  This includes what nodes are being used and by
who, as in which experiments and thus which projects.

NOTE: For a while you might see a few of these warnings,
  *** WARNING: os_setup:
  ***   could not find previous state (rsrcidx=484084) in image_history
  ***   table, won't be able to determine newly allocated nodes
if someone does a swapmod to an experiment that was swapped in
before this commit was installed.  This is because os_setup uses
previous information in the table to determine newly allocated
nodes.  This warning can safely be ignored in this case, and should
go away over time.