Add a "regression" mode to both the old assign_wrapper and the new mapper wrapper. In regression mode, the wrapper/mapper proceeds normally, creating a .vtop file, and then running assign with a fixed seed. If the wrapper and the mapper agree on the .vtop file, then the solution from assign should be identical.
The wrapper/mapper then proceeds normally, reserving resources and
making all the DB changes. Needless to say, this has to be on a
private copy of the database, with all nodes free. Creating that DB
was a tale in its own right.

At completion, call the existing BackupPhysicalState() function that
we use in swapmod, and write all the physical tables we have changed
(just the rows corresponding to the experiment of course). The delete
all that state, and free the nodes.

If everything is working correctly, those physical tables should be
identical when created by the mapper or the wrapper.

Of course, its not quite there yet. I have a few things to fix up
before diff -r produces no results.