Whoa...it's a coding frenzy for mike!
Be more explicit in the final message about deleting a user when you are
just removing that user from a project and not entirely.

Just about pissed my pants when it told me it was deleting a user after
I had confirmed that I REALLY (and REALLY, REALLY) just wanted to delete
them from a project.