Rob and I talked about regression testing last week, and we decided that depending on different versions of assign to find the exact same solution on the same top/ptop, is asking for trouble. And in fact, the new assign is slightly different and the solutions do not match.
So the idea we came up with is to first run the old version, and then
fixnode the results for the input to the new version. assign should
run cleanly and the results will be the same. Then reverse the
situation and run the new version, then fixnode those results for a
run of the old version. Then compare all the results.