Add method to flush and reload the reservation info from the node object.
Method to find out if a physical node has any virtual nodes allocated
to it.

Method to release the bandwidth that a vnode on a shared host is
using. This uses the vinterfaces table to find out what link is using,
and then increments the remaining_bandwidth of the interface_state
table. Tables are locked of course.

Add method to allocate an rtabid, since like ssh ports on remote
nodes, the rtabid is now a shared counter on a sharedhost.

Hmm, a change to the TipServer() method that was for using console on
boss.