Pass interface trunk setting through to the outer emulab. Also pass information about whether the interface needs to trunk in dual mode instead of equal mode. Normally we use equal, but the supervlan created between shared hosts needs to be trunked in dual mode so that "raw" packets are delivered (or else non-encap packets go into the bit bucket).
When tearing down vlans we do not use individual trunk disables, but
do it one block from destroyvlans. This is way way faster then an
XMLRPC call for each individual port.

For trunk ports, no longer use the vinterfaces table, but instead use
the trunk slot of the interfaces table. I removed a Keith Sklower
comment that said how silly that was; he was right, and now its
fixed.