Add -noswapout option to prevent swapout after a botched swapmod. Used by the pool daemon to keep the shared-nodes from getting released.
Add call to new ReserveSharedBandwidth() when doing a swapmod
recovery. See comment in Experiment.pm cvs log.

Use the new snmpit -X option (synchronize tables) instead of vlan
teardown and setup during a swapmod. This is currently enabled
only on the MAINSITE and only for testbed/tbres/emulab-ops until we
get some experience with it.