Add new snmpit -X option (synchronize tables) to use during swapmod, instead of vlan teardown and setup during a swapmod.
This operates by using the vlans and interface_state tables to know
what is onthe switches and what actually need to be changed. Should be
much more efficient.

Note that this requires switch backend changes. I did it for the cisco
backend, which is why tbswap only does it on MAINSITE.