Changed the way that snmpit decides what switches to operate on for the 'tables', 'reset', and 'synctables' commands. Rather than assuming all VLANs live on the One True Experimental Net Stack (unless -S was given), they all now check the database to see which stack they should be operating on.
Changed how stack information is currently stored in the database. The
LAN attribute formerly known as 'stack' is now called 'class', and has
two valid values, 'Experimental' and 'Control'. The idea is to record
whether something is an experimental net or control net LAN, not which
specific stack it's on. The 'stack' column in the VLANs table now
records the actual stack_id on which the VLAN has been created.

This is all in support of testbeds that have multiple experimental-net
stacks, as Utah currently does.