Add '-f' "filemode" option to imagezip for creating image files from regular files (as opposed to from disk partitions).  This is almost identical to rawmode, but also knows how to handle inputs that are not a multiple of the sector size.  Imagezip will pad out the final sector, but uses a special relocation entry (SHORTSECTOR) to tell imageunzip/frisbee how many bytes to write in that final sector.
This is "pretty much" backward compatible.  If an old version of imageunzip
encounters the unknown new relocation entry, it will ignore it, resulting
in imageunzip just writing out the additional pad bytes to the output file.
This may or may not screw over the resulting output file.

Also fixed some lint found by GCC 4.2 on FBSD7.