The 'make' command needs to grab the whole stack, not just switches that contain ports in the new VLAN: we might need to transit switches 'in the middle'.
I'm starting to think that the way we built up list of switches for
the commands is not the right thing to be doing - almost all of the
operations operate on whole stacks now, so we just have extra work to
do to convert stack names into switch names and back. Or, at the very
least, we need  a simpler way to compute (and store) the set of
switches that a VLAN has to transit.