Modified kernel module installation
Kernel modules are now compressed into a tarball in addtion to being
installed into the initramfs.  This way all modules are still available
even if modules are removed from the initramfs to reduce its size.

The final initramfs image is now located in $(TOPDIR) along with the
kernel and modules tarball.