Make sure we delete subnodes of a target node.  A subnode is one whose phys_nodeid is the target node.  In keeping with the Mike school of thinking on hierarchy, we only look for one level of subnode.  If a node does have subnodes, we prompt to make sure the caller really wants to go through with it.
While I was in there, add the -n option to not actually do any DB deletes
or daemon restarts.

While I was in there, add the -S option which first mysqldumps all the
entries we will be deleting.  Maybe this can prevent a catastrophe in the
future.  However, since I didn't make this the default it will likely only
give the caller a greater sense of self-loathing when they do precipitate
the catastrophe ("If only I had used the -S option that Mike the Omniscient
added")