Add the ability to load images on virtnodes.  For now, we just overload the tb-set-node-os command with a second optional argument; if that is present, the first arg is the child OS and the second is the parent OS. We add some new features in ptopgen (OS-parentOSname-childOSname) based off a new table that maps which child OSes can run on which parents, and the right desires get added to match.  We setup the reloads in os_setup along with the parents.  Also needed a new opmode, RELOAD-PCVM, to handle all this.
For now, users only have to specify that their images can run on pcvms, a
special hack for which type the images can run on.  This makes sense in
general since there is no point conditionalizing childOS loading on
hardware type at the moment, but rather on parentOS.  Hopefully this stuff
wiill mostly work on shared nodes too, although we'll have to be more
aggressive on the client side garbage collecting old frisbee'd images for
long-lived shared hosts.

I only made these changes in libvtop, so assign_wrapper folks are left in
the dark.

Currently, the client side supports frisbee.  Only in openvz for now, and
this probably breaks libvnode_xen.pm.  Also in here are some openvz
improvements, like ability to sniff out which network is the public
control net, and which is the fake virtual control net.