Modified rc.linux to support shared swap space
Changed rc.linux to support sharing swap space with FreeBSD, since
newer FreeBSD images want to do this.

On first boot, all swap entries are removed from /etc/fstab to ensure
no invalid entries are present.  Then the root disk is scanned and any
type 0x82 partitions are added to fstab as swap devices.  An empty file
is created in BOOTDIR once complete, indicating to subsequent runs that
this step has already been done.

On each boot, the swap entries in fstab are checked.  If the device
listed is a valid block device and is not currently active, mkswap and
swapon are run to activate it.  Entries that have the "noauto" option
are skipped, as are ones prefixed with `#'.