This version includes heavy duty machinery for parallelizing activities among all switches in a stack.  Does check to see if the required additional perl modules are installed and will revert to standard behavior if not there.
Currently only looking up vlan numbers from all the switches in the stack
is actually parallelized; this alone results in a speed up by a factor of 5
on DETER.

Additional speedups could be gained if we assign all the vlan numbers for
vlans in snmpit -t at the time it is run thereby allowing us to
construct different vlans in different switches in an overlapping way.

We haven't implemented this yet on DETER awaiting synchronizing with the
mothership to make use of VLan->setVlanTag().