Dynamically add lines to inittab for serial consoles
Modified /init to parse /proc/cmdline for console devices.  Each
'console=' parameter is parsed and each serial device (ttyS* or
ttyUSB* for now) specified is used to generate an /etc/inittab entry
that starts a getty on that device.  The baud rate is parsed as well,
but defaults to 115200 if not specifed.  Parity and data/stop bits are
assumed to be N-8-1.

The specified device is also added to /etc/securetty if not already
present.