Fixes for git-cvspush and modified cvsexportcommit
Make git-diff-tree generate more context (8 lines).

Also, keep track of the last commit successfully pushed to CVS and start
there when doing a cvspush if the local repo is up to date.  This should
prevent duplicate commits and allow new commits to be pushed before the
central Git repo is updated from CVS.