Add an Emulab-specific py-m2crypto port.
All we really do is apply one non-standard patch to avoid SSL certificate
mismatches due to "www.emulab.net" vs. "boss.emulab.net".  In older versions
of m2crypto we just patched the installed library files, but the newer versions
get installed as a .egg file, so it is easiest to just patch it while the
port is being built.

I really don't like creating a separate port for this, but the alternative
was to modify the standard port, adding a new option.  That seemed even worse.