One fix and one work around for new (5.10.0) perl warnings about "used once" variables in libtblog_simple and SWIG generated code.
The fix is to libtblog, to get rid of the warning.  The warning from swig
generated code is a swig problem that was supposedly fixed, but apparently
not.  So those warnings remain, and led to the workaround.

The workaround is in shommlists which parses output captured by SUEXEC.
That function captures all output, stdout and stderr, from the command it runs
and presents it in an array.  So if you use it to invoke a perl script that
evokes the "used once" messages, those messages wind up in the array.
showmmlists was not validating that output at all, it just assumed everything
returned was the name of a mail list.  Added some syntactic validation
(aka a regexp) to deal with that.