First phase of adding a special '*' bandwidth to the top file
This value can be specified in place of an integer bandwidth. When
this is done, assign will map the virtual link to *any* physical
link of the right type, regardless of bandwidth. However, it will
adjust the bandwidth used for trunk calculations, etc, based on the
bandwidth of the interfaces selected, so that this doesn't cause us
to overuse interswitch links.

For the time being, it is not legal to use this option with trivial
links or emulated (multiplexed) links, since it's not at all clear
what this would mean.

This commit contains only the data structure and parser changes; the
actual implementation will come soon.