Handle the case of links between a switch and a node - I thought I had disallowed this in the parser, but I hadn't thought about LAN nodes, which typically get assigned to switches. So, handle this case correctly.
We still do *not* handle links between two switches (eg. directly
between two LANs). However, what the semantics of this should be
is totally unclear - do we use the *whole* interswtich bandwidth?
Since we are not likely to hit this in our regular usage, we punt
for now with an error if it happens.

Also, remove the parser check, since we now handle correctly the
case it was trying to disallow.