Bugfix for fixed interfaces
When I moved the code for checking them from resolve_link() into
find_best_link(), I inadvertently dropped the code that was checking
for the 'flipped' flag - in the former location, this was happening
implicitly due to the order links were placed in the resolutions
structure. In the new location, it has to be checked explicitly.

Needs more testing before pushing to main tree.