Major mkextrafs whackage for FreeBSD.
Change mkextrafs so that we can create a filessytem on the second disk.
The command line option was always there, it just didn't work.

Added -l option to make it print out "available" partitions that you could
put a filesystem on.  This uses the new os_find_freedisk() function which
was inspired by Dave's Linux libvnode routine to locate unused space.
In fact, that latter function should get moved up to linux/liblocsetup.pm
and the Linux mkextrafs updated to support -l.