Add node attribute for specifying bootinfo server
Add the bootinfo_server attribute for specifying the bootinfo server the
node should use during the boot process.  If not specified the node will
default to trying the tftp server and dhcp server, in that order.

PXE vendor option 128 is used to communicate this to the NBP. Options
128-254 are available for NBP-specific use according to the PXE spec, so
we're safe using it.