Add revoke option (-r) to grantnodetype script. Does what you think it does.
Change the code that rebuilds nodetypeXpid_permissions so that if
a node is specifically revoked, make sure it is granted to all
other projects. This is kinda gross, but in fact, we really need to
ditch nodetypeXpid_permissions and use the policy tables directly,
but I do not have time to do that.

Remove all that robot lab open/close stuff in libadminctrl. Silly
stuff that is no longer used.