Add ability to choose compression type for initramfs
Add '-c' option to gen_initramfs.sh to choose the compression type.
Currently, it supports the following types: gzip (default), bzip2, lzma,
and none.

The toplevel Makefile now has a variable to specify the compression type
of the initramfs, and passes it to gen_initramfs.sh.  Currently it is
set to lzma by default.