Clear out bogus DB partition info on default reload.
The new d710s have a single-partition image as their default.  As a result,
when they were reloaded, the info for the other partitions (in particular
partition 1) was not being cleared.  So if I were to break my install of
FBSD72-STD on a d710 node and then free it, the reload would not clear
the partition 1 info in the DB.  If someone else came along requesting
FBSD72-STD and got that node assigned, os_load would think that it didn't
need to reload partition 1 and just boot right into the old stale partition.

Now for a "default" reload (i.e., from reload_daemon), a single partition
load will clear all the other partitions.  We don't do this all the time
since a user might want to reload one of the partitions on their allocated
node without clobbering the others.