Fix gitmail's short_refname() to support refs containing slashes
Refs can contain slashes.  For example, the short ref for
'refs/heads/foo/bar' is 'foo/bar'.  Previously, we just chopped off
everything up to and including the last slash character, which doesn't
return the correct short ref in this case.

Call git-rev-parse to shorten the ref for us.  In case it doesn't return
anything (which shouldn't happen), fall back to the full ref.