Finish up user deletion. The big visible change is that when a user is deleted, they still remain in the user table with a status of "archived", but since all the queries in the system now use uid_idx instead of uid, it is safe to reuse a uid since they are no longer ambiguous.
The reason for not deleting users from the users table is so that the
stats records can refer to the original record (who was that person
named "mike"). This is very handy and worth the additional effort it
has taken.

There is no way to ressurect a user, but it would not be hard to add.