No longer install the emulab p5-GD-2.07 port.
We introduced this port or an older version of GD long ago to get around an
incompatibility with perl.  Our perl has since moved on and we no longer
need this.  To be safe, we still install this on 4.x-based installs (though
I am sure many, many other things would break if someone tried to install
Emulab on FreeBSD 4.x now!)