Total reorg of sensorts/nfstrace/nfsdump2 directory.
Previously it contained a copy of the nfsdump utility with local changes.
The actual build of the tool would also happen in the source tree.

This has now been changed to be more like tg/iperf/rude, where it downloads
(in the build tree) the appropriate distribution, applies our patches, and
then builds.  The "patch" includes several new, local files and rather than
literally using patch to create those, there is now a subdir in source tree
called "files" that contains those.  After patch runs, these files are
symlinked into the build tree (symlinking was easier than mucking with the
Makefile to do a VPATH thing).