BUG FIX:  Make parse errors "User" errors again.
The problem is that comit f6d81289c37259a5e21f133a88ab1f9b301f73b2
(Changes from Pat Gunn at CMU: allow "at swapout"...) changed the exit
status of parse.proxy with this line:
  -    exec("nice -15 $parser @ARGV $nsfile");
  +    exec("nice -15 $parser @ARGV $nsfile|tee $tempdir/output.txt");
so parse-ns was not detecting that it failed.  It then processed to
run xmlconvert which detected that the ns parse failed.

The causes two problems, 1) my error handling code depends on the fact
that parse.proxy failed in order to determine it was a user error and
2) it causes the "Syntax Check" option to always report correct syntax
even when it shouldn't.